requirements.txt
pyTelegramBotAPI==4.14.0
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
boto3==1.34.0
botocore==1.34.0 